Review comments - Naming of keystore - Other cosmetic changes
